+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #3 - created passport entity
 - #4 - entity BankCard was created, integrated into gazon-service and tested
 - #6 - added pagination for Example
+- #7 - entities Product and ProductImage were created and integrated
 
 ## [1.0.0]